Moved cat out of support, removed support
