+++ examples/cat.c
