+++ examples/cat.h
