--- examples/support/README
@@ -1,2 +1,0 @@
-Support code for the examples. Code that's not relevant to the principle each
-example demonstrates is largely offloaded to this directory.
